rename: 전체 구조 위치 변경에 따른 수정
