+++ src/main/java/com/example/springrider/domain/review/controller/ReviewController.java
@@ -1,9 +1,10 @@
 package com.example.springrider.domain.review.controller;
 
-import com.example.springrider.domain.common.response.ApiResponse;
+
 import com.example.springrider.domain.review.dto.CreateReviewRequestDto;
 import com.example.springrider.domain.review.dto.ReviewResponseDto;
 import com.example.springrider.domain.review.service.ReviewService;
+import com.example.springrider.global.response.ApiResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PathVariable;
